% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mta_lines}
\alias{mta_lines}
\title{NYC MTA subway lines}
\format{The mta_lines dataframe contains 23 observations on 1 variable.
\describe{
  \item{NYC MTA lines}{MTA line number}
}}
\source{
\url{http://www.mta.info}
}
\usage{
mta_lines
}
\description{
A dataset containing the 23 MTA lines running in NYC.
}
\keyword{datasets}
